refactor: make test framework more readable
